requirements.txt
cython==3.0.10
imgaug==0.4.0
ipython==8.12.3
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.9.0.80
pillow==10.0.0
protobuf==3.19.6
pycocotools~=2.0.7
pandas==1.2.0
pyyaml==6.0.1
scikit-image==0.16.2
scipy==1.7.3
tensorflow==2.6.0
einops~=0.8.0
tqdm~=4.66.4
open3d~=0.18.0

